Updated type of argument

+++ energy-rs/src/lib.rs
@@ -55,10 +55,10 @@ pub mod energy {
         }
     }
     
-    fn read_msr(raw: RawFd, r : u32) -> u64 {
+    fn read_msr(raw: RawFd, r : i64) -> u64 {
         
         let mut buf = [0u8;8];
-        let _ = pread(raw, &mut buf, r as i64);
+        let _ = pread(raw, &mut buf, r);
         let r = unsafe { mem::transmute::<[u8; 8], u64>(buf) };
         r
     }
